search_sublet result fixed
Now the search results are displayed in 3 columns if the screen is in large size, 2 for middle and 1 for small/xs.  each card is not having a fixed width anymore, the card's width will change according to the browser's size. Also feel free to edit card-body for each searched result if you think you can make it more reasonable and neat.